Add cliff changes
